requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pip==23.0.1
pytz==2023.3
setuptools==66.0.0
sqlparse==0.4.4
wheel==0.38.4
